+++ src/sessions/base-session-handler.ts
@@ -14,10 +14,10 @@ type ExtendedHTMLAudioElement = HTMLAudioElement & {
 
 export default abstract class BaseSessionHandler {
   removePendingSessionDelay = 1000;
+  disabled = true;
+  abstract sessionType: SessionTypes;
 
   constructor (protected sdk: PureCloudWebrtcSdk, protected sessionManager: SessionManager) { }
-
-  abstract sessionType: SessionTypes;
 
   abstract shouldHandleSessionByJid (jid: string): boolean;